리뷰 : 별점 정수-> double 형으로 변경하여 상세 벌졈 줄수있도록 수정
효과/가성비/ 향·맛·/복용 /편의성

+++ lib/presentation/shopping/widgets/producrt_normal_review.dart
@@ -339,8 +339,8 @@ class _ReviewStats {
     final satisfiedCount = reviews.where((r) => r.isRecommend == 'y').length;
     final dissatisfiedCount = reviews.where((r) => r.isRecommend == 'n').length;
 
-    double avgScore(int Function(ReviewModel) pick) {
-      return reviews.map((r) => pick(r).toDouble()).reduce((a, b) => a + b) /
+    double avgScore(double Function(ReviewModel) pick) {
+      return reviews.map((r) => pick(r)).reduce((a, b) => a + b) /
           reviews.length;
     }
